+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.18.1
+
+* added `BinaryParser.get_remaining_bytes`
+* added `BinaryParser.get_remaining_bytes_count`